unify dataConverters
