+++ web/src/components/organisms/Project/ModelsMenu/hooks.ts
@@ -5,6 +5,7 @@ import Notification from "@reearth-cms/components/atoms/Notification";
 import { Model } from "@reearth-cms/components/molecules/Model/types";
 import { Group, ModelFormValues } from "@reearth-cms/components/molecules/Schema/types";
 import { fromGraphQLModel } from "@reearth-cms/components/organisms/DataConverters/model";
+import { fromGraphQLGroup } from "@reearth-cms/components/organisms/DataConverters/schema";
 import {
   useGetModelsQuery,
   useGetGroupsQuery,
@@ -18,7 +19,6 @@ import {
 } from "@reearth-cms/gql/graphql-client-api";
 import { useT } from "@reearth-cms/i18n";
 import { useModel, useWorkspace, useProject } from "@reearth-cms/state";
-import { fromGraphQLGroup } from "@reearth-cms/utils/values";
 
 type Params = {
   modelId?: string;
